* #10: Moved form logic from various auth view components into separate form components for better separation of concerns.     * ForgotPassword.tsx:         - component that contains the form definition and logic for password reset previously defined in ForgotPasswordView.     * RegisterView.tsx:         - Deleted Register form definition and logic.         - Imports RegisterForm component instead. (not inside this commit)     * LoginView.tsx:         - Deleted Login form definition and logic.         - Imports LoginForm component instead. (not inside this commit)     * ForgotPasswordView:         - Deleted Forgot password form definition and logic.         - Imports ForgotPassword component instead.
